+++ packages/core/src/components/MultiSelect/Options/Options.styled.ts
@@ -40,9 +40,9 @@ export const ChipArea = styled.div<WithThemeProp>`
     }
 `;
 
-export const OptionsWrapper = styled.div<WithThemeProp>`
+export const OptionsWrapper = styled.div<WithThemeProp & { size: 'S' | 'M' }>`
     position: absolute;
-    top: 5.6rem;
+    top: ${({ size }) => (size === 'S' ? '4rem' : '5.6rem')};
     box-shadow: ${({ theme }) => `0px 2px 8px ${theme.colors.grey[400]}`};
     background-color: ${({ theme }) => theme.colors.white};
     box-sizing: border-box;