+++ packages/core/src/components/SingleSelect/Options/types.ts
@@ -1,8 +1,9 @@
 import { HTMLProps, WithThemeProp } from '@medly-components/utils';
 import { Option } from '../types';
 
-export interface OptionsProps extends HTMLProps<HTMLUListElement> {
+export interface OptionsProps extends Omit<HTMLProps<HTMLUListElement>, 'size'> {
     id?: string;
+    size?: 'S' | 'M';
     options: Option[];
     variant: 'outlined' | 'filled';
     isNested?: boolean;
@@ -11,6 +12,7 @@ export interface OptionsProps extends HTMLProps<HTMLUListElement> {
 }
 
 export interface StyledOptionsProps extends WithThemeProp {
+    size?: 'S' | 'M';
     isNested?: boolean;
     variant: 'outlined' | 'filled';
 }