+++ packages/core/src/components/TextField/TextField.stories.tsx
@@ -6,6 +6,7 @@ import { TextField } from './TextField';
 import { Props } from './types';
 
 export const variants: Props['variant'][] = ['filled', 'outlined'];
+export const sizes: Props['size'][] = ['S', 'M'];
 
 export const ThemeInterface: React.FC<TextFieldTheme> = () => null;
 ThemeInterface.defaultProps = {