+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.1](https://github.com/medly/medly-components/compare/@medly-components/forms@1.8.0...@medly-components/forms@1.8.1) (2020-08-27)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+# [1.8.0](https://github.com/medly/medly-components/compare/@medly-components/forms@1.7.1...@medly-components/forms@1.8.0) (2020-08-27)
+
+
+### Features
+
+* add small size for textfield, singleselect, multiselect component ([7da9d6b](https://github.com/medly/medly-components/commit/7da9d6b89e1038dfbcf58a0e2332184305b0dc51))
+
+
+
+
+
 ## [1.7.1](https://github.com/medly/medly-components/compare/@medly-components/forms@1.7.0...@medly-components/forms@1.7.1) (2020-08-25)
 
 **Note:** Version bump only for package @medly-components/forms