+++ packages/loaders/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.15](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.0.14...@medly-components/loaders@1.0.15) (2020-08-27)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
 ## [1.0.14](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.0.13...@medly-components/loaders@1.0.14) (2020-08-25)
 
 **Note:** Version bump only for package @medly-components/loaders